BACKEND/teacher_management/created_teachers/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
email-validator
